Create Get_Vertex_Thickness.sh
This is a bash shell script for running vertex-wise cortical thickness analysis on two groups of subjects for comparison in Freesurfer v5.x. In order to use the script you need: Freesurfer v5.X, and a directory structure set of up for pre-processed subject data (i.e., recon-all). The script provides detailed instructions on how to edit the appropriate variables within the script, and instructions on how to create needed text files for Freesurfer to discern what features to use for estimating the parameters of a GLM for comparing two groups with varying cortical thickness.